исправления 1 правки Base -- нужно base, пакеты именуются в таком формате.Перепроверь, плиз, формат нейминга всех пакетов. для всех тестов на логин нужно вынести в before создание тестового курьера, а в after его нужно удалять В параметризованных тестах для аннотации @Parameterized.Parameters лучше использовать аргумент name: @Parameterized.Parameters(name = "Тестовые данные: {0} {1}"), где {0}, {1} - индексы параметров. Это повысит информативность проверки.

+++ src/test/java/CourierLoginTest.java
@@ -1,6 +1,7 @@
-import Courier.*;
+import courier.*;
 import io.restassured.response.ValidatableResponse;
 import org.junit.After;
+import org.junit.Before;
 import org.junit.Test;
 
 import static org.hamcrest.Matchers.equalTo;
@@ -8,28 +9,39 @@ import static org.hamcrest.Matchers.equalTo;
 public class CourierLoginTest {
     private CourierGenerator generator = new CourierGenerator();
     private CourierMethod client = new CourierMethod();
+    private Courier courier;
     private int courierId;
+
+    @Before
+    public void setup() {
+        // Создаем курьера перед каждым тестом
+        courier = generator.random();
+        client.create(courier);
+    }
 
     @After
     public void deleteCourier() {
-        if (courierId > 0) {
-            client.delete(courierId);
+        // Удаляем курьера после каждого теста
+        if (courier != null) {
+            ValidatableResponse response = client.login(courier);
+            courierId = client.checkLogin(response);
+            if (courierId > 0) {
+                client.delete(courierId);
+            }
         }
     }
 
     @Test
     public void successLoginTest() {
-        Courier courier = generator.random();
-        client.create(courier);
         ValidatableResponse response = client.login(courier);
         courierId = client.checkLogin(response);
         assert courierId != 0;
     }
 
     @Test
-    public void incorrectLogin() {
-        Courier courier = generator.random();
-        client.login(courier)
+    public void incorrectLoginTest() {
+        Courier invalidCourier = generator.random(); // Создаем нового курьера, который не был зарегистрирован
+        client.login(invalidCourier)
                 .assertThat()
                 .statusCode(404)
                 .body("message", equalTo("Учетная запись не найдена"));
